feat: implement ApplyFormContent component for housing application
